# api-contract-v1.yaml
openapi: 3.1.0
info:
  title: Event Face Matching API
  version: 1.0.0
  description: |
    Phase 1 contract — single studio, no billing. Multi-tenant structures are
    present but only one studio row exists.

    ## Who owns what
    Backend (routes, DB, worker) is implemented by the studio owner.
    Frontend (three React apps) is implemented against this spec via a mock server.
    Neither side blocks the other. When the real server is live, only the base URL changes.

    ## Two separate auth contexts — do not mix them
    - **Staff** (studio owners, photographers, platform admins) use
      `Authorization: Bearer <access_token>`.
    - **Guests** have no account. They use `X-Guest-Session: <session_token>`,
      scoped to exactly one event, and can reach only the `/g/*` routes.

    Keeping these in different headers is deliberate: it makes it structurally
    hard to accidentally accept a guest token on a staff route.

    ## photos.status state machine
    ```
    pending ──► uploaded ──► processing ──► done
                                 │
                                 └────────► failed ──► (retry) ──► processing
    ```
    - `pending`    row created, presigned URL issued, bytes not yet confirmed
    - `uploaded`   client confirmed the PUT succeeded; job enqueued
    - `processing` a worker has claimed the job
    - `done`       faces extracted, embedded, stored, matched against guests
    - `failed`     gave up after 5 attempts; `error_code` explains why

...

...
servers:
...
security:
  - staffAuth: []
paths:
...
  /g/{qr_token}:
    parameters:
      - name: qr_token
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [guest]
      summary: Public event info behind a QR code
      description: |
        The first request a guest makes. Returns only what is safe to show an
        unauthenticated stranger holding the token: event name, date, and the
        branding to render. No photos, no guest list, no studio internals.
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GuestEventInfo' }
        '404': { $ref: '#/components/responses/NotFound' }
  /g/{qr_token}/session:
    parameters:
      - name: qr_token
        in: path
        required: true
        schema: { type: string }
    post:
      tags: [guest]
      summary: Record consent and open a guest session
      description: |
        Must be called before the camera is opened. `consented` must be `true` —
        there is no path to a session without explicit consent, because the
        selfie that follows is biometric data under the DPDP Act.

        The returned `session_token` goes in `X-Guest-Session` on every
        subsequent `/g/*` call.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [consented]
              properties:
                consented:
                  type: boolean
                  const: true
      responses:
        '201':
          description: Session opened
          content:
            application/json:
              schema:
                type: object
                required: [session_token, expires_at, face_retention_days, face_deletion_date]
                properties:
                  session_token: { type: string }
                  expires_at: { type: [string, 'null'], format: date-time }
                  face_retention_days:
                    type: integer
                    description: |
                      Snapshotted onto this session at the moment of consent, from
                      the event, which snapshotted it from the studio's negotiated
                      setting. Never a live lookup: what a guest agreed to cannot
                      be changed afterwards by editing a studio setting.

                      Not a tier feature and not purchasable. Capped platform-wide
                      by MAX_FACE_RETENTION_DAYS.
                  face_deletion_date:
                    type: string
                    format: date
                    description: |
                      The actual calendar date this guest's selfie is deleted. The
                      consent screen shows this rather than a duration, so the
                      promise is specific and checkable rather than a general claim
                      about policy.
  /g/selfie:
    post:
      tags: [guest]
      summary: Register the guest's face and run the initial search
      description: |
        Runs **once** per guest. Detects, quality-checks, embeds and stores the
        selfie, then searches every face indexed so far for this event and
        writes the matches.

        From here on, matching happens in the worker as new photos arrive. The
        gallery never calls this again.

        Expects a JPEG under 2MB, already compressed client-side.
      security:
        - guestAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [selfie]
              properties:
                selfie: { type: string, format: binary }
      responses:
        '200':
          description: Face registered
          content:
            application/json:
              schema:
                type: object
                required: [matched_count]
                properties:
                  matched_count:
                    type: integer
                    description: Photos matched from those already indexed.
        '422':
          description: |
            The selfie was unusable. `code` tells the frontend which specific
            retake instruction to show — these are the most common failure in
            the whole product, so each needs its own human message.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                no_face:
                  value: { error: { code: NO_FACE_DETECTED, message: We could not find a face in that photo. } }
                multiple_faces:
                  value: { error: { code: MULTIPLE_FACES, message: Please take a selfie with just yourself in frame. } }
                too_small:
                  value: { error: { code: FACE_TOO_SMALL, message: Please hold the camera closer. } }
                too_blurry:
                  value: { error: { code: FACE_TOO_BLURRY, message: That photo was too blurry. Hold still and try again. } }
  /g/photos:
    get:
      tags: [guest]
      summary: The guest's matched photos
      description: |
        Polled roughly every 20 seconds while the event runs.

        This is a plain database read — no model inference, no external API,
        no per-call cost. That is the whole reason the architecture works.

        Pass `after` (the previous `latest_cursor`) to fetch only newly matched
        photos. The frontend uses this to show "8 new photos" rather than
        reshuffling the grid under the guest's thumb.
      security:
        - guestAuth: []
      parameters:
        - name: after
          in: query
          schema: { type: string }
          description: Return only matches newer than this cursor.
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items, next_cursor, latest_cursor, total_count]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/GuestPhoto' }
                  next_cursor: { type: [string, 'null'] }
                  latest_cursor:
                    type: string
                    description: Pass as `after` on the next poll.
                  total_count: { type: integer }
  /g/session:
    delete:
      tags: [guest]
      summary: Delete the guest's face data and matches
      description: |
        The DPDP deletion path. Removes the embedding, the session and all
        match rows immediately — not on a schedule.

        Does **not** delete the studio's photos; the guest has no rights over
        those. It only removes the link between this person and them.
      security:
        - guestAuth: []
      responses:
        '204':
          description: Deleted

  # ───────────────────────────── members ─────────────────────────────
...
components:
  securitySchemes:
    staffAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Studio owners, photographers and platform admins.
    guestAuth:
      type: apiKey
      in: header
      name: X-Guest-Session
      description: Opaque session token, scoped to one event, valid for `/g/*` only.
  parameters:
...
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
  responses:
...
    NotFound:
      description: |
        Not found — also returned when the resource exists but belongs to
        another studio. Never distinguish these two cases; doing so leaks
        the existence of other tenants' data.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable identifier. Frontend switches on this, never on `message`.
            message:
              type: string
              description: Human-readable text, safe to show a user.
            details:
              type: object
              additionalProperties: true
...
    GuestEventInfo:
      type: object
      required: [event_name, event_date, branding, accepting_guests, face_retention_days]
      properties:
        event_name: { type: string }
        event_date: { type: string, format: date }
        accepting_guests:
          type: boolean
          description: False once the event has ended.
        face_retention_days:
          type: integer
          description: |
            Shown before consent, so a guest knows what they are agreeing to
            before the camera opens rather than after.
        branding:
          type: object
          required: [mode]
          properties:
            mode: { type: string, enum: [platform, studio] }
            studio_name: { type: [string, 'null'] }
            logo_url: { type: [string, 'null'], format: uri }
            brand_color: { type: [string, 'null'] }
    GuestPhoto:
      type: object
      required: [photo_id, thumbnail_url, full_url, matched_at]
      properties:
        photo_id: { type: string, format: uuid }
        thumbnail_url: { type: string, format: uri, description: Presigned GET, ~1 hour. }
        full_url: { type: string, format: uri, description: Presigned GET, ~1 hour. }
        width: { type: integer }
        height: { type: integer }
        matched_at: { type: string, format: date-time }